requirements.txt
asgiref==3.5.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==4.0.4
django-cors-headers==3.12.0
django-heroku==0.0.0
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
whitenoise==6.1.0
psycopg2
